requirements.txt
streamlit
PyPDF2
transformers
sentence-transformers
faiss-cpu
numpy
torch
